Adding plot records and estimated pool netspace
This adds a record keeping table for plots submitted during partial submission.  We record only new plots and refresh each plot when a new partial of that plot comes in.

We also delete plots after a certain time when they go stale.  The target time should be around the time you expect each plot to send a new partial.  This way we only keep valid and active plots in the records for each farmer.

We can use the plot records for various things including updating frontend applications for the user.