Update specification to allow multi-pool servers
`target_puzzle_hash` parameter for was partially documented in `GET /login` documentation and not introduced yet in the implementation. This PR makes the documentation consistent.

It was not said anywhere, I think, but this, when implemented, will help pools to be hosted easily, meaning that a pool operator will be able to outsource his operation to a SaaS which will then handle most of the grunt work, at least for the `GET /login` part which doesn't handle any real money. Specifically I think this allows multiple pools to be hosted on the same domain and it allows a pool to change to a different domain. It also prevents attacks where one pool tries to steal an authentication cookie and use it to manage the account of a different pool the given user is a member of. It should at least be documented well. Preliminary look on `get_pool()` suggests that proper target_puzzle_hash verification should not be hard, but lets put that in a separate PR.